backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
PyPDF2==3.0.1
pycryptodome==3.19.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pdfplumber==0.10.3 